fix: code review — placeholder collision, API hygiene, CI/CD, docs (#5)
* fix: code review fixes — placeholder collision, API hygiene, CI/CD, docs

- Single-pass code-block extraction with PUA-bracketed markers (U+E000)
  fixes placeholder collision and O(n²) regression for identical blocks
- `default_length_counter` returns `usize`; `Chunk.length` is `usize`
- `#[non_exhaustive]` on `Chunk` (output type) for forward compatibility
- Empty input now returns [] (symmetric with whitespace-only input)
- Pin MSRV to Rust 1.80 in Cargo.toml (required for LazyLock)
- Drop musl branches and isMusl() from Node loader; document supported platforms
- CI matrix expanded to ubuntu/macos/windows; coverage gate at 99%
- Release workflow: validate (fmt+clippy+tests+dry-run) before bump/tag/push
- Dependabot automerge: use fetch-metadata@v2 instead of hand-rolled regex
- Codecov targets lowered to 99% (accommodates llvm-cov phantom line)
- Docs: clarify length contract, title semantics, supported Markdown formats


* perf: single-pass restore_code_placeholders, write! in extract_code_blocks


* style: cargo fmt


* fix: coverage tests for restore edge cases; drop CI OS matrix


* fix: address Copilot review — deny position, doc, debug_assert, platform error, header PUA, release.yml


* fix: musl detection, assert! for length cast, node: prefix in loader


---------

+++ crate/src/lib.rs
@@ -1,8 +1,18 @@
+#![deny(clippy::all)]
 //! Heading-aware, token-budgeted semantic chunker for Markdown.
 //!
 //! Splits a Markdown document by heading hierarchy and merges/splits chunks
 //! to stay within a character budget. Designed for RAG pipelines and embedding
 //! workflows where section context matters.
+//!
+//! ## Supported Markdown
+//!
+//! - **Headers:** ATX only (`# H1` through `###### H6`). Setext headers
+//!   (`=====`/`-----` underlines) and headings with 7+ `#` are not recognized.
+//! - **Code blocks:** backtick-fenced (` ``` `) and inline (`` ` `` … `` ` ``).
+//!   Tilde fences (`~~~`) and 4-space-indented code are **not** protected —
+//!   `#` lines inside them are parsed as headers. Switch to backtick fences
+//!   to avoid mis-splits.
 //!
 //! ## Algorithm
 //!
@@ -42,8 +52,8 @@ pub use types::{Chunk, ChunkOptions};
 pub fn chunk(text: &str, options: Option<ChunkOptions>) -> Vec<Chunk> {
     let opts = options.unwrap_or_default();
 
-    let min_length = opts.min_length.unwrap_or(512);
-    let max_length = opts.max_length.unwrap_or(3072);
+    let min_length = opts.min_length.unwrap_or(512) as usize;
+    let max_length = opts.max_length.unwrap_or(3072) as usize;
     let phase = opts.phase.unwrap_or(3);
     let title = opts.title.as_deref();
